backend/inpainting-ai/requirements.txt
# API / FastAPI stack
fastapi==0.116.1
starlette==0.47.3
uvicorn==0.35.0
starlette-context==0.4.0
python-multipart==0.0.20

# Pydantic
pydantic==2.11.9
pydantic-settings==2.10.1

# Imagen / procesamiento numérico
numpy==1.24.3
Pillow==11.3.0
opencv-python-headless==4.11.0.86

# OCR
paddlepaddle==3.2.0
paddleocr==3.2.0

# Inpainting / Lama Cleaner stack - UPDATED VERSIONS
# lama-cleaner==1.2.5  # Updated to newer version
# diffusers==0.16.1    # Updated to compatible version
transformers==4.27.4
# huggingface-hub==0.34.4  # Updated to compatible version
# diffusers==0.14.0
# huggingface-hub==0.12.1

# Torch stack
torch==2.8.0
torchvision==0.23.0

# AWS etc
boto3==1.40.30
botocore==1.40.30
aws-secretsmanager-caching==1.1.3

# Utilidades
requests==2.28.2
structlog==25.4.0
protobuf==4.25.8
# scipy==1.15.3
loguru==0.7.2

# Utilidades
scipy==1.9.3
rich==13.9.4
addict==2.4.0
einops==0.8.0
timm==0.9.12
omegaconf==2.3.0
scikit-image==0.19.3
pyyaml==6.0.2
easyocr==1.7.1

